requirements.txt
requests>=2.28.1
pandas>=1.5.3
matplotlib>=3.7.2
chembl-webresource-client>=0.10.1
pyyaml>=6.0
pytest>=7.4.0
myst-parser>=3.0.1
mdit_py_plugins>=0.4.2
markdown-it-py>=3.0.0
sphinx>=7,<9
sphinx-rtd-theme>=3.0.1
myst-parser>=3
docutils>=0.21
